mmlu ready for test
